# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://127.0.0.1:3000"
      }
    ],
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    },
...
    "paths": {
...
      "/train": {
        "get": {
          "tags": [
            "Train"
          ],
          "summary": "/train",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {}
              }
            }
          }
        },
        "post": {
          "tags": [
            "Train"
          ],
          "summary": "/train",
          "requestBody": {
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "YESGO-101"
                    },
                    "label": {
                      "type": "string",
                      "example": "Train YESGO Paris St Lazarre - Caen"
                    },
                    "start_station": {
                      "type": "string",
                      "example": "Paris SL"
                    },
                    "end_station": {
                      "type": "string",
                      "example": "Gare de Caen"
                    },
                    "time_of_departure": {
                      "type": "string",
                      "example": "06:50"
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {}
              }
            }
          }
        }
      },
...
      "/trainstation": {
        "get": {
          "tags": [
            "TrainStation"
          ],
          "summary": "/trainstation",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {}
              }
            }
          }
        },
        "post": {
          "tags": [
            "TrainStation"
          ],
          "summary": "/trainstation",
          "requestBody": {
            "content": {
              "application/x-www-form-urlencoded": {
                "schema": {
                  "properties": {
                    "label": {
                      "type": "string",
                      "example": "Gare Rouen Rive Droite test"
                    },
                    "imageGare": {
                      "type": "string",
                      "example": "https://www.architecture-art-deco.fr/photos/rouen/medium/gare-art-deco02.jpg"
                    },
                    "open_hour": {
                      "type": "string",
                      "example": "06:00"
                    },
                    "close_hour": {
                      "type": "string",
                      "example": "22:00"
                    }
                  }
                }
              }
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {}
              }
            }
          }
        }
      },
...
    }
  }